+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added support for the `stack run` command.
 
+* Improved automatic opening of generated Haddocks.
+
 ## Hasky Stack 0.8.0
 
 * Make commands like “build → bench” and “build → test” only propose targets